[SWOZO-27] added persistence of VMS

+++ orchestrator/src/main/java/com/swozo/orchestrator/cloud/resources/vm/TimedVMProvider.java
@@ -8,7 +8,7 @@ import java.util.concurrent.CompletableFuture;
 public interface TimedVMProvider {
     CompletableFuture<VMResourceDetails> createInstance(Psm psm) throws InterruptedException, VMOperationFailed;
 
-    CompletableFuture<Void> deleteInstance(int internalResourceId) throws InterruptedException, VMOperationFailed;
+    CompletableFuture<Void> deleteInstance(long internalResourceId) throws InterruptedException, VMOperationFailed;
 
     int getVMCreationTime(Psm psm);
 }
